Reoganize tests to match module reorg
