+++ test/iterators.js
@@ -3,101 +3,6 @@ var tr = require('../'),
     iter = require('../').iterators,
     test = require('tape')
 
-test('iterate array', function(t){
-  var arr, iterator, idx
-
-  idx = 0
-  arr = [1,2,3]
-  iterator = tr.iterator(arr)
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  idx = 0
-  arr = [2]
-  iterator = tr.iterator(arr)
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  idx = 0
-  arr = []
-  iterator = tr.iterator(arr)
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  t.end()
-})
-
-test('iterate string', function(t){
-  var arr, iterator, idx
-
-  idx = 0
-  arr = ['1','2','3']
-  iterator = tr.iterator('123')
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  idx = 0
-  arr = ['2']
-  iterator = tr.iterator('2')
-  t.deepEquals({value: arr[idx++], done: false}, iterator.next())
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  idx = 0
-  iterator = tr.iterator('')
-  t.ok(iterator.next().done)
-  t.ok(iterator.next().done)
-
-  t.end()
-})
-
-test('iterate object', function(t){
-  var obj, arr
-
-  obj = {a:1, b:2, c:3}
-  arr = [['a', 1],['b', 2],['c', 3]]
-
-  t.deepEqual(iter.toArray(obj).sort(), arr)
-  t.deepEqual(iter.toArray({}), [])
-
-  t.end()
-})
-
-
-test('iterate fn', function(t){
-  var fn, iterator, start
-
-  function count(init){
-    var cnt = init
-    return function(){
-      return cnt++
-    }
-  }
-
-  start = 0
-  iterator = tr.iterator(count(start))
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-
-  start = 10
-  iterator = tr.iterator(count(start))
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-  t.deepEquals({value: start++, done: false}, iterator.next())
-
-  t.end()
-})
-
 test('range', function(t){
   var fn, iterator, start, iterable
