+++ test/string.js
